Replace int with size_t to describe the type of the variables in more detail.
[[originally from CVS; cvs2svn created svn r4273]]

+++ client/attribute.h
@@ -47,45 +47,45 @@ enum attr_tile {
 void attribute_init(void);
 void attribute_flush(void);
 void attribute_restore(void);
-void attribute_set(int key, int id, int x, int y, int data_length,
+void attribute_set(int key, int id, int x, int y, size_t data_length,
 		   const void *const data);
-int attribute_get(int key, int id, int x, int y, int max_data_length,
+size_t attribute_get(int key, int id, int x, int y, size_t max_data_length,
 		  void *data);
 
 /*
  * Special methods for units.
  */
-void attr_unit_set(enum attr_unit what, int unit_id, int data_length,
+void attr_unit_set(enum attr_unit what, int unit_id, size_t data_length,
 		   const void *const data);
-int attr_unit_get(enum attr_unit what, int unit_id, int max_data_length,
+size_t attr_unit_get(enum attr_unit what, int unit_id, size_t max_data_length,
 		   void *data);
 void attr_unit_set_int(enum attr_unit what, int unit_id, int data);
-int attr_unit_get_int(enum attr_unit what, int unit_id, int *data);
+size_t attr_unit_get_int(enum attr_unit what, int unit_id, int *data);
 
 
 /*
  * Special methods for cities.
  */
-void attr_city_set(enum attr_city what, int city_id, int data_length,
+void attr_city_set(enum attr_city what, int city_id, size_t data_length,
 		   const void *const data);
-int attr_city_get(enum attr_city what, int city_id, int max_data_length,
+size_t attr_city_get(enum attr_city what, int city_id, size_t max_data_length,
 		   void *data);
 void attr_city_set_int(enum attr_city what, int city_id, int data);
-int attr_city_get_int(enum attr_city what, int city_id, int *data);
+size_t attr_city_get_int(enum attr_city what, int city_id, int *data);
 
 /*
  * Special methods for players.
  */
-void attr_player_set(enum attr_player what, int player_id, int data_length,
+void attr_player_set(enum attr_player what, int player_id, size_t data_length,
 		     const void *const data);
-int attr_player_get(enum attr_player what, int player_id,
-		    int max_data_length, void *data);
+size_t attr_player_get(enum attr_player what, int player_id,
+		    size_t max_data_length, void *data);
 
 /*
  * Special methods for tiles.
  */
-void attr_tile_set(enum attr_tile what, int x, int y, int data_length,
+void attr_tile_set(enum attr_tile what, int x, int y, size_t data_length,
 		   const void *const data);
-int attr_tile_get(enum attr_tile what, int x, int y, int max_data_length,
+size_t attr_tile_get(enum attr_tile what, int x, int y, size_t max_data_length,
 		  void *data);
 #endif
